allow dictionary-filtering logic to detect multiples of a letter
So if a letter appears multiple times in a guess (and one of them is
in the correct place, but the other one isn't), the filtering logic will
not wrongly filter all words containing the letter.

Instead, the filtering logic will get rid of all words with that letter
in the spot where we know it isn't present.